Tidying up, delinting, and pruning. (#108)
* Moves components around.

* Fixes a rogue warning.

* Removes linting issues.

* Removes dead code.
