+++ packages/reactotron-app/App/Commands/ApiResponseCommand.js
@@ -1,10 +1,10 @@
 import React, { Component, PropTypes } from 'react'
-import Command from '../Command'
-import ObjectTree from '../../Shared/ObjectTree'
+import Command from '../Shared/Command'
+import ObjectTree from '../Shared/ObjectTree'
 import { dotPath, isWithin } from 'ramdasauce'
 import { toUpper } from 'ramda'
-import Colors from '../../Theme/Colors'
-import makeTable from '../../Shared/MakeTable'
+import Colors from '../Theme/Colors'
+import makeTable from '../Shared/MakeTable'
 
 const Styles = {
   container: {
