+++ packages/reactotron-app/App/Foundation/PageStreaming.js
@@ -1,33 +1,32 @@
 import React, { Component } from 'react'
-import { observer, inject } from 'mobx-react'
+import { observer } from 'mobx-react'
 import AppStyles from '../Theme/AppStyles'
 import Colors from '../Theme/Colors'
+import CommandList from '../Shared/CommandList'
 
 const Styles = {
   container: {
     ...AppStyles.Layout.vbox,
     backgroundColor: Colors.screen,
-    paddingLeft: 200,
-    paddingRight: 150,
+    paddingLeft: 0,
+    paddingRight: 0,
     paddingTop: 0,
     paddingBottom: 0,
-    overflowY: 'scroll',
-    marginRight: 4
+    overflowY: 'scroll'
   }
 }
 
-@inject('session')
 @observer
-class PageNetworking extends Component {
+class PageStreaming extends Component {
 
   render () {
     return (
       <div style={Styles.container}>
-        Networking
+        <CommandList />
       </div>
     )
   }
 
 }
 
-export default PageNetworking
+export default PageStreaming
