+++ packages/reactotron-app/App/Foundation/VisualRoot.js
@@ -1,13 +1,10 @@
 import React, { Component } from 'react'
 import Header from './Header'
 import Page from './Page'
-import PageLogging from '../Logging/PageLogging'
-import PageState from '../State/PageState'
-import PageNetworking from '../Network/PageNetworking'
 import Footer from './Footer'
 import Colors from '../Theme/Colors'
 import AppStyles from '../Theme/AppStyles'
-import PageStreaming from '../Streaming/PageStreaming'
+import PageStreaming from './PageStreaming'
 import ReactTooltip from 'react-tooltip'
 import SampleModal from '../Dialogs/SampleModal'
 
@@ -26,17 +23,6 @@ export default class VisualRoot extends Component {
     return (
       <div style={Styles.container}>
         <Header />
-        {/*
-        <Page tabId='logging'>
-          <PageLogging />
-        </Page>
-        <Page tabId='state'>
-          <PageState />
-        </Page>
-        <Page tabId='network'>
-          <PageNetworking />
-        </Page>
-        */}
         <Page tabId='streaming'>
           <PageStreaming />
         </Page>
