+++ packages/reactotron-app/App/Shared/CommandIcon.js
